# clockee-ui/src/gen/models/PageResponseAdminProductDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminProductDTO } from './AdminProductDTO';
export type PageResponseAdminProductDTO = {
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    data?: Array<AdminProductDTO>;
};